drop Typelike SAFE_MODIFIED_COPY_METHODS. implement Hashlike#select, #reject, #compact, #merge. implement Arraylike #select, #reject, #compact. no longer implement #transform_values to return a modified copy.
this fixes an issue where Hashlike#select and #reject returned a modified copy (a SchemaInstanceBase or JSON::HashNode) but the block variables were from the underlying hash, instead of the receiver.

transform_values isn't practical to return a modified copy because putting block result into the typelike object is ... weird.