no longer checking schema.describes_hash? / schema.describes_array? on SchemaInstanceBase initialize to decide to extend with SchemaInstanceBaseHash / SchemaInstanceBaseArray.
I think it doesn't make sense for this to check the schema. although it is a bit unexpected for a SIB instance for a schema the descirbes an array to have hash methods, it is better that SIB just wrap the node per its type.

more often the schema doesn't specify its type, and it is not desirable for scorpio to fail to decorate an instance as a hash due to a more open ended schema.