Util.as_json uses object's #as_json if defined
disregarding a defined #as_json when an object responds to #to_hash or whatever was not so good.
eliminates check for Base, since Base responds to #as_json.
also drops the call to Util.as_json on the result of object.as_json - its as_json could return something not actually jsonifiable but distrusting that is unnecessary.