SchemaAncestorNode#jsi_anchor_subschemas return a SchemaSet
a SchemaSet is typically a set of schemas that all describe some instance, which this is not, and that is the reason it was a Set and not a SchemaSet before.
but this Set, like SchemaSet, should compare_by_identity. since Set#compare_by_identity is not universally available, and since there's no particular reason a SchemaSet should always contain schemas that describe the same instance, this will be a SchemaSet.